+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/service/DaoOnlyRelationService.java
@@ -8,20 +8,31 @@ import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
 import com.dwarfeng.subgrade.stack.log.LogLevel;
 import com.dwarfeng.subgrade.stack.service.RelationService;
 
+import javax.annotation.Nonnull;
+
 /**
  * 仅通过数据访问层实现的关系服务。
- * <p>该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。</p>
+ *
+ * <p>
+ * 该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。
  *
  * @author DwArFeng
  * @since 0.2.3-beta
  */
 public class DaoOnlyRelationService<PK extends Key, CK extends Key> implements RelationService<PK, CK> {
 
+    @Nonnull
     private RelationDao<PK, CK> dao;
+    @Nonnull
     private ServiceExceptionMapper sem;
+    @Nonnull
     private LogLevel exceptionLogLevel;
 
-    public DaoOnlyRelationService(RelationDao<PK, CK> dao, ServiceExceptionMapper sem, LogLevel exceptionLogLevel) {
+    public DaoOnlyRelationService(
+            @Nonnull RelationDao<PK, CK> dao,
+            @Nonnull ServiceExceptionMapper sem,
+            @Nonnull LogLevel exceptionLogLevel
+    ) {
         this.dao = dao;
         this.sem = sem;
         this.exceptionLogLevel = exceptionLogLevel;
@@ -54,27 +65,30 @@ public class DaoOnlyRelationService<PK extends Key, CK extends Key> implements R
         }
     }
 
+    @Nonnull
     public RelationDao<PK, CK> getDao() {
         return dao;
     }
 
-    public void setDao(RelationDao<PK, CK> dao) {
+    public void setDao(@Nonnull RelationDao<PK, CK> dao) {
         this.dao = dao;
     }
 
+    @Nonnull
     public ServiceExceptionMapper getSem() {
         return sem;
     }
 
-    public void setSem(ServiceExceptionMapper sem) {
+    public void setSem(@Nonnull ServiceExceptionMapper sem) {
         this.sem = sem;
     }
 
+    @Nonnull
     public LogLevel getExceptionLogLevel() {
         return exceptionLogLevel;
     }
 
-    public void setExceptionLogLevel(LogLevel exceptionLogLevel) {
+    public void setExceptionLogLevel(@Nonnull LogLevel exceptionLogLevel) {
         this.exceptionLogLevel = exceptionLogLevel;
     }