+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/service/DaoOnlyWriteService.java
@@ -15,23 +15,30 @@ import java.util.Objects;
 
 /**
  * 仅通过数据访问层实现的写入服务。
- * <p>该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。</p>
+ *
+ * <p>
+ * 该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。
  *
  * @author DwArFeng
  * @since 1.1.0
  */
 public class DaoOnlyWriteService<K extends Key, E extends Entity<K>> implements WriteService<E> {
 
+    @Nonnull
     private WriteDao<E> dao;
+    @Nonnull
     private KeyFetcher<K> keyFetcher;
+    @Nonnull
     private ServiceExceptionMapper sem;
+    @Nonnull
     private LogLevel exceptionLogLevel;
 
     public DaoOnlyWriteService(
             @Nonnull WriteDao<E> dao,
             @Nonnull KeyFetcher<K> keyFetcher,
             @Nonnull ServiceExceptionMapper sem,
-            @Nonnull LogLevel exceptionLogLevel) {
+            @Nonnull LogLevel exceptionLogLevel
+    ) {
         this.dao = dao;
         this.keyFetcher = keyFetcher;
         this.sem = sem;
@@ -50,6 +57,7 @@ public class DaoOnlyWriteService<K extends Key, E extends Entity<K>> implements 
         }
     }
 
+    @Nonnull
     public WriteDao<E> getDao() {
         return dao;
     }
@@ -58,6 +66,7 @@ public class DaoOnlyWriteService<K extends Key, E extends Entity<K>> implements 
         this.dao = dao;
     }
 
+    @Nonnull
     public KeyFetcher<K> getKeyFetcher() {
         return keyFetcher;
     }
@@ -66,6 +75,7 @@ public class DaoOnlyWriteService<K extends Key, E extends Entity<K>> implements 
         this.keyFetcher = keyFetcher;
     }
 
+    @Nonnull
     public ServiceExceptionMapper getSem() {
         return sem;
     }
@@ -74,11 +84,22 @@ public class DaoOnlyWriteService<K extends Key, E extends Entity<K>> implements 
         this.sem = sem;
     }
 
+    @Nonnull
     public LogLevel getExceptionLogLevel() {
         return exceptionLogLevel;
     }
 
     public void setExceptionLogLevel(@Nonnull LogLevel exceptionLogLevel) {
         this.exceptionLogLevel = exceptionLogLevel;
+    }
+
+    @Override
+    public String toString() {
+        return "DaoOnlyWriteService{" +
+                "dao=" + dao +
+                ", keyFetcher=" + keyFetcher +
+                ", sem=" + sem +
+                ", exceptionLogLevel=" + exceptionLogLevel +
+                '}';
     }
 }