+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/service/GeneralEntireLookupService.java
@@ -12,23 +12,33 @@ import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
 import com.dwarfeng.subgrade.stack.log.LogLevel;
 import com.dwarfeng.subgrade.stack.service.EntireLookupService;
 
+import javax.annotation.Nonnegative;
 import javax.annotation.Nonnull;
 import java.util.List;
 
 /**
  * 通用的全体实体查询服务。
- * <p>该类同时使用数据访问层和缓存实现实体的查询方法。</p>
- * <p>该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。</p>
+ *
+ * <p>
+ * 该类同时使用数据访问层和缓存实现实体的查询方法。
+ *
+ * <p>
+ * 该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。
  *
  * @author DwArFeng
  * @since 0.0.3-beta
  */
 public class GeneralEntireLookupService<E extends Entity<?>> implements EntireLookupService<E> {
 
+    @Nonnull
     private EntireLookupDao<E> dao;
+    @Nonnull
     private ListCache<E> cache;
+    @Nonnull
     private ServiceExceptionMapper sem;
+    @Nonnull
     private LogLevel exceptionLogLevel;
+    @Nonnegative
     private long cacheTimeout;
 
     public GeneralEntireLookupService(
@@ -36,7 +46,7 @@ public class GeneralEntireLookupService<E extends Entity<?>> implements EntireLo
             @Nonnull ListCache<E> cache,
             @Nonnull ServiceExceptionMapper sem,
             @Nonnull LogLevel exceptionLogLevel,
-            long cacheTimeout
+            @Nonnegative long cacheTimeout
     ) {
         this.dao = dao;
         this.cache = cache;
@@ -141,6 +151,7 @@ public class GeneralEntireLookupService<E extends Entity<?>> implements EntireLo
         }
     }
 
+    @Nonnull
     public EntireLookupDao<E> getDao() {
         return dao;
     }
@@ -149,6 +160,7 @@ public class GeneralEntireLookupService<E extends Entity<?>> implements EntireLo
         this.dao = dao;
     }
 
+    @Nonnull
     public ListCache<E> getCache() {
         return cache;
     }
@@ -157,6 +169,7 @@ public class GeneralEntireLookupService<E extends Entity<?>> implements EntireLo
         this.cache = cache;
     }
 
+    @Nonnull
     public ServiceExceptionMapper getSem() {
         return sem;
     }
@@ -165,6 +178,7 @@ public class GeneralEntireLookupService<E extends Entity<?>> implements EntireLo
         this.sem = sem;
     }
 
+    @Nonnull
     public LogLevel getExceptionLogLevel() {
         return exceptionLogLevel;
     }
@@ -179,5 +193,16 @@ public class GeneralEntireLookupService<E extends Entity<?>> implements EntireLo
 
     public void setCacheTimeout(long cacheTimeout) {
         this.cacheTimeout = cacheTimeout;
+    }
+
+    @Override
+    public String toString() {
+        return "GeneralEntireLookupService{" +
+                "dao=" + dao +
+                ", cache=" + cache +
+                ", sem=" + sem +
+                ", exceptionLogLevel=" + exceptionLogLevel +
+                ", cacheTimeout=" + cacheTimeout +
+                '}';
     }
 }